MAJOR Menu Implementations
-Added Scene transitions
-Added data transfer across Scenes
-Menu now correctly implements data into Dojo Scene & game
-Reordered Hair Materials on Dojo Scene
-Added SceneManager & scripts for scene transitions
-Edited the way outfit variation works
-Edited the way Blendshapes are applied
-Added option to use data from MainMenu in Dojo Scene

+++ Assets/Scripts/Menu/GrowingText.cs
@@ -7,7 +7,7 @@ using UnityEngine.EventSystems;
 
 public class GrowingText : MonoBehaviour
 {
-    public int largeScaleSize = 3;
+    public float largeScaleSize = 3;
     float normalScaleSize;
     GameObject currentSelected;
     GameObject eventSystem;
